requirements.txt
torch
torchvision
torchaudio

tensorflow-macos
tensorflow-metal

scipy
scikit-learn
pandas
pyod
pyyaml
matplotlib
pickle

